% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{calculate_topology}
\alias{calculate_topology}
\title{Calculate Network Topology}
\usage{
calculate_topology(g, output_dir = NULL, prefix = "Network")
}
\arguments{
\item{g}{igraph object.}

\item{output_dir}{Directory to save topology files (optional).}

\item{prefix}{Prefix for output files.}
}
\description{
Calculate Network Topology
}
