% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model.R
\name{fit_sncm}
\alias{fit_sncm}
\title{Fit Neutral Community Model}
\usage{
fit_sncm(spp, pool = NULL, taxon = NULL)
}
\arguments{
\item{spp}{Abundance matrix (rows are samples, columns are OTUs).}

\item{pool}{Metacommunity abundance (optional).}

\item{taxon}{Taxonomy information (optional).}
}
\description{
Fit Neutral Community Model
}
