% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_network.R
\name{generate_random_networks}
\alias{generate_random_networks}
\title{Generate Random Networks}
\usage{
generate_random_networks(n, e, reps = 1000, output_file = NULL)
}
\arguments{
\item{n}{Number of nodes.}

\item{e}{Number of edges.}

\item{reps}{Number of random networks to generate, default 1000.}

\item{output_file}{Output file path for summary (optional).}
}
\description{
Generate Random Networks
}
