% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model.R
\name{plot_sncm_fit}
\alias{plot_sncm_fit}
\title{Plot Neutral Community Model Fit}
\usage{
plot_sncm_fit(spp_out, fill = NULL, title = NULL)
}
\arguments{
\item{spp_out}{Output object from fit_sncm.}

\item{fill}{Variable to fill points with, default "fit_class".}

\item{title}{Plot title.}
}
\description{
Plot Neutral Community Model Fit
}
