% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icamp_analysis.R
\name{run_icamp}
\alias{run_icamp}
\title{Run iCAMP Analysis}
\usage{
run_icamp(
  comm,
  tree,
  treat,
  env = NULL,
  clas,
  prefix = "iCAMP_Test",
  save_dir = tempdir(),
  rand = 1000,
  nworker = 4,
  memory_G = 50,
  bin_size_limit = 24
)
}
\arguments{
\item{comm}{Community abundance matrix (rows are samples, columns are OTUs).}

\item{tree}{Phylogenetic tree object (phylo).}

\item{treat}{Group information (data frame).}

\item{env}{Environmental variables (data frame, optional).}

\item{clas}{Taxonomy information (data frame).}

\item{prefix}{Prefix for output files.}

\item{save_dir}{Directory to save outputs and intermediate files.}

\item{rand}{Randomization times, default 1000.}

\item{nworker}{Number of threads, default 4.}

\item{memory_G}{Memory limit in GB, default 50.}

\item{bin_size_limit}{Minimum bin size, default 24.}
}
\value{
List containing iCAMP results (icres and icbin).
}
\description{
Run iCAMP Analysis
}
